+++ src/components/blocks/RoleFilterCardLayout.tsx
@@ -2,7 +2,6 @@ import { X } from "lucide-react";
 import type { ReactNode } from "react";
 
 interface RoleFilterCardLayoutProps {
-	guid: string;
 	onDelete: () => void;
 	header: ReactNode;
 	children: ReactNode;
@@ -12,16 +11,12 @@ interface RoleFilterCardLayoutProps {
  * Role Filter Card の基本レイアウトコンポーネント (Stateless)
  */
 export function RoleFilterCardLayout({
-	guid,
 	onDelete,
 	header,
 	children,
 }: RoleFilterCardLayoutProps) {
 	return (
-		<div
-			data-guid={guid}
-			className="bg-white shadow rounded-lg p-4 border border-gray-200 flex flex-col gap-3 relative"
-		>
+		<div className="bg-white shadow rounded-lg p-4 border border-gray-200 flex flex-col gap-3 relative">
 			<button
 				type="button"
 				onClick={onDelete}
