+++ src/feature/rolefilter/RoleFilterCardHeader.tsx
@@ -1,0 +1,67 @@
+import { Plus } from "lucide-react";
+import { postRoleFilterUpdate, roleFilterMetaData } from "../../logics/api";
+import { PostExRAssignOps } from "../../type";
+import { useStore } from "../../useStore";
+
+interface RoleFilterCardHeaderProp {
+	guid: string;
+	assignNum: number;
+	excludeRoleIds: number[];
+}
+
+export function RoleFilterCardHeader({
+	guid,
+	assignNum,
+	excludeRoleIds,
+}: RoleFilterCardHeaderProp) {
+	const openBlockDialog = useStore((state) => state.openBlockDialog);
+	const addRoleToFilter = useStore((state) => state.addRoleToFilter);
+
+	const onOpenRoleSelect = () => {
+		openBlockDialog({
+			type: "roleSelect",
+			title: "役職の追加",
+			searchQuery: "",
+			excludeRoleIds: excludeRoleIds,
+			selectedRoleIds: [],
+			lastClickedId: null,
+			onSelect: async (roleIds: number[]) => {
+				try {
+					for (const roleId of roleIds) {
+						await postRoleFilterUpdate({
+							Op: PostExRAssignOps.FilterRoleAdd,
+							FilterId: guid,
+							MapRoleId: roleId,
+						});
+
+						const roleName =
+							(roleFilterMetaData.NormalRoleId[roleId] as string) ||
+							(roleFilterMetaData.CombinationId[roleId] as string) ||
+							(roleFilterMetaData.GhostRoleId[roleId] as string) ||
+							"Unknown Role";
+
+						addRoleToFilter(guid, roleId, roleName);
+					}
+				} catch (error) {
+					console.error("Failed to add role to filter:", error);
+				}
+			},
+		});
+	};
+
+	return (
+		<>
+			<span className="text-sm font-semibold text-gray-700">
+				AssignNum: {assignNum}
+			</span>
+			<button
+				type="button"
+				onClick={onOpenRoleSelect}
+				className="mt-1 self-start inline-flex items-center px-2 py-1 border border-transparent text-xs font-medium rounded text-indigo-700 bg-indigo-100 hover:bg-indigo-200 focus:outline-none"
+			>
+				<Plus size={12} className="mr-1" aria-hidden="true" />
+				役職を追加
+			</button>
+		</>
+	);
+}
